fix: move methods into utils file DATATR-65
